[master] Added includes of new BIND9 compatibility headers, updated util/bind.sh
    Merges in rt48072.

+++ server/dhcpv6.c
@@ -1034,7 +1034,8 @@ void check_pool6_threshold(struct reply_state *reply,
 				  shared_name,
 				  inet_ntop(AF_INET6, &lease->addr,
 					    tmp_addr, sizeof(tmp_addr)),
-				  used, count);
+				  (long long unsigned)(used),
+				  (long long unsigned)(count));
 		}
 		return;
 	}
@@ -1066,7 +1067,8 @@ void check_pool6_threshold(struct reply_state *reply,
 		  "address: %s; high threshold %d%% %llu/%llu.",
 		  shared_name,
 		  inet_ntop(AF_INET6, &lease->addr, tmp_addr, sizeof(tmp_addr)),
-		  poolhigh, used, count);
+		  poolhigh, (long long unsigned)(used),
+		  (long long unsigned)(count));
 
 	/* handle the low threshold now, if we don't
 	 * have one we default to 0. */
@@ -1436,12 +1438,15 @@ pick_v6_address(struct reply_state *reply)
 		log_debug("Unable to pick client address: "
 			  "no addresses available  - shared network %s: "
 			  " 2^64-1 < total, %llu active,  %llu abandoned",
-			  shared_name, active - abandoned, abandoned);
+			  shared_name, (long long unsigned)(active - abandoned),
+			  (long long unsigned)(abandoned));
 	} else {
 		log_debug("Unable to pick client address: "
 			  "no addresses available  - shared network %s: "
 			  "%llu total, %llu active,  %llu abandoned",
-			  shared_name, total, active - abandoned, abandoned);
+			  shared_name, (long long unsigned)(total),
+			  (long long unsigned)(active - abandoned),
+		          (long long unsigned)(abandoned));
 	}
 
 	return ISC_R_NORESOURCES;
